Explicitly passed datatype to np.*_like functions.

+++ pmutt/mixture/__init__.py
@@ -48,7 +48,8 @@ def _get_mix_quantity(misc_models, method_name, raise_error=True,
         return np.array([default_value])
 
     # Calculate contribution from mixing models if any
-    mix_quantity = np.full_like(a=misc_models, fill_value=default_value)
+    mix_quantity = np.full_like(a=misc_models, fill_value=default_value,
+                                dtype=np.double)
     for i, mix_model in enumerate(misc_models):
         if mix_model is None:
             continue
